+++ pybamm/models/submodels/active_material/base_active_material.py
@@ -83,10 +83,11 @@ class BaseModel(pybamm.BaseSubModel):
             # R_n, R_p. For a size distribution, calculate the area-weighted
             # mean using the distribution instead. Then the surface area is
             # calculated the same way
-            if self.options["particle size"] == "single":
+            domain_options = getattr(self.options, domain)
+            if domain_options["particle size"] == "single":
                 R = self.phase_param.R
                 R_dim = self.phase_param.R_dimensional
-            elif self.options["particle size"] == "distribution":
+            elif domain_options["particle size"] == "distribution":
                 if self.domain == "negative":
                     R_ = pybamm.standard_spatial_vars.R_n
                 elif self.domain == "positive":